review: address comments

+++ crates/rust-client/src/errors.rs
@@ -135,9 +135,9 @@ pub enum ClientError {
     #[error("client initialization error: {0}")]
     ClientInitializationError(String),
     #[error("note tags limit exceeded (max {0})")]
-    NoteTagsLimitExceeded(usize),
+    NoteTagsLimitExceeded(u32),
     #[error("accounts limit exceeded (max {0})")]
-    AccountsLimitExceeded(usize),
+    AccountsLimitExceeded(u32),
     #[error("unsupported authentication scheme ID: {0}")]
     UnsupportedAuthSchemeId(u8),
     #[error("account error is not full: {0}")]
