+++ crates/rust-client/src/lib.rs
@@ -396,7 +396,7 @@ impl<AUTH> Client<AUTH> {
     /// Checks if the note tag limit has been exceeded.
     pub async fn check_note_tag_limit(&self) -> Result<(), ClientError> {
         let limits = self.rpc_api.get_rpc_limits().await?;
-        if self.store.get_unique_note_tags().await?.len() >= limits.note_tags_limit {
+        if self.store.get_unique_note_tags().await?.len() >= limits.note_tags_limit as usize {
             return Err(ClientError::NoteTagsLimitExceeded(limits.note_tags_limit));
         }
         Ok(())
@@ -405,7 +405,7 @@ impl<AUTH> Client<AUTH> {
     /// Checks if the account limit has been exceeded.
     pub async fn check_account_limit(&self) -> Result<(), ClientError> {
         let limits = self.rpc_api.get_rpc_limits().await?;
-        if self.store.get_account_ids().await?.len() >= limits.account_ids_limit {
+        if self.store.get_account_ids().await?.len() >= limits.account_ids_limit as usize {
             return Err(ClientError::AccountsLimitExceeded(limits.account_ids_limit));
         }
         Ok(())
